fix: Created entities namespace for Placetype
To mirror the Tfl api namespacing, I've created a new Entities namespace for certain types. It seems that the Entities namespace holds generic models that are general across different endpoints.
