+++ app/graphql/types/stop_point_type.rb
@@ -2,7 +2,7 @@ module Types
   class StopPointType < Types::BaseObject
     field :accessibilitySummary, String, null: true
     field :additionalProperties, [Types::AdditionalPropertiesType, null: true], null: true
-    field :children, [Types::PlaceType, null: true], null: true
+    field :children, [Types::Tfl::Entities::PlaceType, null: true], null: true
     field :childrenUrls, [String, null: true], null: true
     field :commonName, String, null: true
     field :distance, Float, null: true
